ThePhasesOfCraftship/2_best_practice_first/testingBasics/exercises/1_Red_Green_Refactor/1_2_Palindrome/src/index.ts
export interface IPalindromeChecker {
  isAPalindrome: (word: string) => boolean
  getReverseWord: (word: string) => string
  getWordWithoutSpaces: (word: string) => string
}

export default class PalindromeChecker implements IPalindromeChecker {
  
  getReverseWord(word: string): string {
    return word.split('').reverse().join('')
  }
  
  getWordWithoutSpaces(word: string): string {
    return word.replace(/\s/g, '')
  }
  
  isAPalindrome(word: string) {
    const reverseWord = this.getReverseWord(this.getWordWithoutSpaces(word)).toLowerCase()
    return this.getWordWithoutSpaces(word).toLowerCase() == reverseWord
  }
}
